# docs/schemas/openapi/admin.spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amp Admin API",
...
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/datasets/{namespace}/{name}/versions/{revision}/deploy": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Handler for the `POST /datasets/{namespace}/{name}/versions/{revision}/deploy` endpoint",
        "description": "Schedules a data extraction job for the specified dataset revision.\n\n## Path Parameters\n- `namespace`: Dataset namespace\n- `name`: Dataset name\n- `revision`: Revision (version, hash, \"latest\", or \"dev\")\n\n## Request Body\n- `end_block`: End block configuration (null for continuous, \"latest\", number, or negative offset)\n- `parallelism`: Number of parallel workers (default: 1, only for raw datasets)\n- `worker_id`: Optional worker selector (exact ID or glob pattern)\n- `verify`: Enable cryptographic verification of EVM block data (default: false)\n- `retry_strategy`: Optional retry strategy controlling how failed jobs are retried\n\n## Response\n- **202 Accepted**: Job successfully scheduled\n- **400 Bad Request**: Invalid path parameters or request body\n- **404 Not Found**: Dataset or revision not found\n- **500 Internal Server Error**: Database or scheduler error\n\n## Error Codes\n- `INVALID_PATH`: Invalid path parameters (namespace, name, or revision)\n- `INVALID_BODY`: Invalid request body (malformed JSON or missing required fields)\n- `DATASET_NOT_FOUND`: The specified dataset or revision does not exist\n- `LIST_VERSION_TAGS_ERROR`: Failed to list version tags from dataset store\n- `RESOLVE_REVISION_ERROR`: Failed to resolve revision to manifest hash\n- `GET_DATASET_ERROR`: Failed to load dataset from store\n- `WORKER_NOT_AVAILABLE`: Specified worker not found or inactive\n- `SCHEDULER_ERROR`: Failed to schedule extraction job\n\n## Behavior\nThis endpoint schedules a data extraction job for a dataset:\n1. Resolves the revision to find the corresponding version tag\n2. Loads the full dataset configuration from the dataset store\n3. Schedules an extraction job with the specified parameters\n4. Returns job ID for tracking\n\nThe revision parameter supports four types:\n- Semantic version (e.g., \"1.2.3\") - uses that specific version\n- \"latest\" - resolves to the highest semantic version\n- \"dev\" - resolves to the development version tag\n- Manifest hash (SHA256 hash) - finds the version that points to this hash\n\nJobs are executed asynchronously by worker nodes. Use the returned job ID\nto track progress via the jobs endpoints.",
        "operationId": "deploy_dataset",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Dataset namespace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Dataset name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "in": "path",
            "description": "Revision (version, hash, latest, or dev)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Job successfully scheduled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeployResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid parameters)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset or revision not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeployRequest": {
        "type": "object",
        "description": "Request for deploying a dataset",
        "properties": {
          "end_block": {
            "$ref": "#/components/schemas/EndBlock",
            "description": "The end block configuration for the deployment\n\nSupports multiple modes:\n- `null` or omitted: Continuous dumping (never stops)\n- `\"latest\"`: Stop at the latest available block\n- `<number>`: Stop at specific block number (e.g., `1000000`)\n- `<negative number>`: Stop N blocks before latest (e.g., `-100` means latest - 100)\n\nIf not specified, defaults to continuous mode."
          },
          "parallelism": {
            "type": "integer",
            "format": "int32",
            "description": "Number of parallel workers to run\n\nEach worker will be responsible for an equal number of blocks.\nFor example, if extracting blocks 0-10,000,000 with parallelism=10,\neach worker will handle a contiguous section of 1 million blocks.\n\nOnly applicable to raw datasets (EVM RPC, Firehose, etc.).\nDerived datasets ignore this parameter.\n\nDefaults to 1 if not specified.",
            "minimum": 0
          },
          "retry_strategy": {
            "type": [
              "object",
              "null"
            ],
            "description": "Optional retry strategy controlling how failed jobs are retried.\n\nSupports three strategies:\n- `none`: Never retry — fail immediately on first error.\n- `bounded`: Retry up to `max_attempts` times with configurable backoff.\n- `unless_stopped`: Retry indefinitely until the job is explicitly stopped.\n\nIf not specified, the scheduler applies its default strategy."
          },
          "verify": {
            "type": "boolean",
            "description": "Enable cryptographic verification of EVM block data during extraction.\n\nWhen enabled, verifies block hashes, transaction roots, and receipt roots\nbefore writing data to storage. Only applicable to EVM raw datasets.\nVerification failures are retryable errors.\n\nDefaults to false if not specified."
          },
          "worker_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional worker selector - either an exact worker ID or a glob pattern.\n\nExamples:\n- `\"worker-node-0\"` - assigns to the specific worker with ID \"worker-node-0\"\n- `\"worker-eth-*\"` - randomly selects from workers matching the pattern\n\nIf not specified, a worker will be selected randomly from all available workers.\nIf a glob pattern is provided, one matching worker will be randomly selected\nThe worker must be active (has sent heartbeats recently) for the deployment to succeed."
          }
        }
      },
      "DeployResponse": {
        "type": "object",
        "description": "Response for deploy operation",
        "required": [
          "job_id"
        ],
        "properties": {
          "job_id": {
            "type": "integer",
            "format": "int64",
            "description": "The ID of the scheduled dump job (64-bit integer)"
          }
        }
      },
      "EndBlock": {
        "type": [
          "string",
          "null"
        ],
        "description": "End block configuration for API requests.\n\nDetermines when the dump process should stop extracting blocks.\nAccepts the following values:\n\n- `null` (or omitted): Continuous dumping - never stops, keeps extracting new blocks as they arrive\n- `\"latest\"`: Stop at the latest available block at the time the dump starts\n- A positive number as a string (e.g., `\"1000000\"`): Stop at the specified absolute block number\n- A negative number as a string (e.g., `\"-100\"`): Stop at (latest block - N), useful for staying N blocks behind the chain tip"
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by the API\n\nThis struct represents error information returned in HTTP error responses.\nIt provides structured error details including a machine-readable error code\nand human-readable message.\n\n## Error Code Conventions\n- Error codes use SCREAMING_SNAKE_CASE (e.g., `DATASET_NOT_FOUND`)\n- Codes are stable and can be relied upon programmatically\n- Messages may change and should only be used for display/logging\n\n## Example JSON Response\n```json\n{\n  \"error_code\": \"DATASET_NOT_FOUND\",\n  \"error_message\": \"dataset 'eth_mainnet' version '1.0.0' not found\"\n}\n```",
        "required": [
          "error_code",
          "error_message"
        ],
        "properties": {
          "error_code": {
            "type": "string",
            "description": "Machine-readable error code in SCREAMING_SNAKE_CASE format\n\nError codes are stable across API versions and should be used\nfor programmatic error handling. Examples: `INVALID_SELECTOR`,\n`DATASET_NOT_FOUND`, `METADATA_DB_ERROR`"
          },
          "error_message": {
            "type": "string",
            "description": "Human-readable error message\n\nMessages provide detailed context about the error but may change\nover time. Use `error_code` for programmatic decisions."
          }
        }
      },
...
    }
  },
...
}